lib: moved result to a standalone module
`engine::result` -> `result`
