+++ crates/rsonpath-lib/src/result.rs
